Add inline Lesson name editing
In the practice view for Lessons we allow editing the name of the lesson
using an inline form. We use the same, now generalized somewhat
inline validation reporting mechanism we used for Section form

- Extract SectionsHelper methods for error field generation to
  ApplicationsHelper
- Extract errors turbo stream view to a shared views directory to be
  able to use it from the Section and Lesson controllers
- Add `InlineValidatable` concern to allow models to define which fields
  they'd like to use inline, extracting the previously existing list
  from SectionsHelper and adding it to the Lesson model as well
- Mark the forms as controlled by the newly introduced Error stimulus
  controller for inline validation
- Add routes, actions and turbo_stream responses to:
  - Replace Lesson name with inline form
  - Respond on success with the same name view (could be extracted to
    its own partial for reuse)
