+++ app/assets/stylesheets/lesson.css
@@ -25,6 +25,13 @@
   }
 }
 
+.title {
+  display: flex;
+  flex-flow: row nowrap;
+  align-items: baseline;
+  gap: 10px;
+}
+
 .video-card {
   display: grid;
