+++ app/controllers/lessons_controller.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class LessonsController < ApplicationController
-  before_action :set_lesson, only: [:show, :edit, :update, :destroy]
+  before_action :set_lesson, only: [:show, :edit, :update, :destroy, :edit_name_inline, :update_inline]
 
   def index
     @lessons = Lesson.includes(:instrument, :sections).where(user_id: current_user.id).page(params[:page])
@@ -46,6 +46,22 @@ class LessonsController < ApplicationController
     respond_to do |format|
       format.html { redirect_to(lessons_url, notice: t(".success")) }
       format.turbo_stream { flash.now[:notice] = t(".success") }
+    end
+  end
+
+  def edit_name_inline
+    respond_to do |format|
+      format.turbo_stream
+    end
+  end
+
+  def update_inline
+    respond_to do |format|
+      if @lesson.update(lesson_params)
+        format.turbo_stream
+      else
+        format.turbo_stream { render("shared/errors", status: :unprocessable_entity, locals: { object: @lesson }) }
+      end
     end
   end
