+++ app/models/concerns/inline_validatable.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module InlineValidatable
+  extend ActiveSupport::Concern
+
+  def inline_validation_fields
+  end
+end
