--- app/views/sections/errors.turbo_stream.erb
@@ -1,5 +1,0 @@
-<% inline_validate.each do |field| %>
-  <%= turbo_stream.replace "#{field}_errors" do %>
-    <%= field_error(field, @section.errors) %>
-  <% end %>
-<% end %>
